requirements.txt
Flask==0.12.2
flask-mongoengine==0.9.3
factory_boy==2.9.2
jsonschema==2.6.0
pytest==3.3.1
pytest-flask==0.10.0
pytest-cov==2.5.1
tox==2.9.1